+++ RedPandaIDE/mainwindow.cpp
@@ -7121,3 +7121,12 @@ void MainWindow::on_actionFilesView_Hide_Non_Support_Files_toggled(bool /* arg1 
     }
 }
 
+
+void MainWindow::on_actionToggle_Block_Comment_triggered()
+{
+    Editor * editor = mEditorList->getEditor();
+    if (editor != NULL ) {
+        editor->toggleBlockComment();
+    }
+}
+